requirements.txt
anyio==3.7.1
Authlib==1.2.1
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
load-dotenv==0.1.0
pycparser==2.21
python-dotenv==1.0.0
sniffio==1.3.0
